requirements.txt
# 个人智能体 - 依赖（在 personal_agent 环境中安装：pip install -r requirements.txt）
# 页面分析 / API 调用 / Webhook / 配置
requests>=2.28.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
